--- dev/20190415/k2works/fizzbuzz/src/main/java/org/hiroshima_arc/domain/model/FizzBuzzData.java
@@ -1,24 +1,0 @@
-package org.hiroshima_arc.domain.model;
-
-import java.util.List;
-
-public class FizzBuzzData {
-    private String _value;
-    private List<String> _values;
-
-    public String getValue() {
-        return _value;
-    }
-
-    public void setValue(String _value) {
-        this._value = _value;
-    }
-
-    public List<String> getValues() {
-        return _values;
-    }
-
-    public void setValues(List<String> _values) {
-        this._values = _values;
-    }
-}
